+++ src/layout/header/Header.tsx
@@ -7,6 +7,7 @@ import './Header.css'
 import cn from 'classnames'
 import ToggleThemeButton from './ToggleThemeButton'
 import HeaderLinks from './HeaderLinks/HeaderLinks'
+import { LanguageToggle } from 'src/layout/header/LanguageToggle'
 
 const { Header } = Layout
 
@@ -17,6 +18,7 @@ const MainHeader = () => {
     <Header className={cn('main-header', { dark: isDarkTheme })}>
       <MenuOutlined onClick={() => setDrawerOpen(true)} className="hideOnDesktop" />
       <div style={{ flex: 1 }}>&nbsp;</div>
+      <LanguageToggle />
       <ToggleThemeButton toggleTheme={toggleTheme} isDarkTheme={isDarkTheme} />
       <HeaderLinks />
     </Header>
